+++ src/torch_dimensions/mixers/__init__.py
@@ -3,6 +3,7 @@ from torch_dimensions.mixers.base import Mixer
 from torch_dimensions.mixers.conv import ConvMixer, TCNMixer, axis_receptive_field
 from torch_dimensions.mixers.rnn import GRUMixer, LSTMMixer
 from torch_dimensions.mixers.ssm import MambaMixer, S4DMixer, S4Mixer
+from torch_dimensions.mixers.upstream import UpstreamMambaMixer, UpstreamS4DMixer, UpstreamS4Mixer
 
 __all__ = [
     "AttentionMixer",
@@ -14,5 +15,8 @@ __all__ = [
     "S4DMixer",
     "S4Mixer",
     "TCNMixer",
+    "UpstreamMambaMixer",
+    "UpstreamS4DMixer",
+    "UpstreamS4Mixer",
     "axis_receptive_field",
 ]